@@ -1,15 +1,19 @@
 {
   "name": "blackbonjour/service-manager",
   "type": "library",
-  "description": "",
-  "keywords": [],
+  "description": "Small, lightweight and factory-biased service manager based on the PSR-11 standard.",
+  "keywords": [
+    "Service Manager",
+    "PSR-11",
+    "Container"
+  ],
   "homepage": "https://github.com/BlackBonjour/service-manager",
   "license": "MIT",
   "authors": [
     {
       "name": "Erick Dyck",
       "email": "info@erickdyck.de",
-      "homepage": "http://www.blackbonjour.de",
+      "homepage": "https://www.erickdyck.de",
       "role": "Developer"
     }
   ],
@@ -19,12 +23,11 @@
     "source": "https://github.com/BlackBonjour/service-manager"
   },
   "require": {
-    "php": ">=7.3",
-
-    "psr/container": "^1.0.0"
+    "php": "^8.1",
+    "psr/container": "^2.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "9.*"
+    "phpunit/phpunit": "^10.0"
   },
   "autoload": {
     "psr-4": {
