@@ -7,46 +7,40 @@ namespace BlackBonjour\ServiceManager;
 use ArrayAccess;
 use BlackBonjour\ServiceManager\AbstractFactory\AbstractFactoryInterface;
 use BlackBonjour\ServiceManager\Exception\ContainerException;
+use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 use Throwable;
-use TypeError;
 
 use function array_key_exists;
 use function is_callable;
 use function is_string;
 
 /**
- * @author    Erick Dyck <info@erickdyck.de>
- * @since     13.05.2019
- * @copyright Copyright (c) 2019 Erick Dyck
+ * @author Erick Dyck <info@erickdyck.de>
+ * @since  13.05.2019
  */
 class ServiceManager implements ArrayAccess, ContainerInterface
 {
-    /** @var AbstractFactoryInterface[] */
-    private $abstractFactories;
-
-    /** @var FactoryInterface[]|callable[]|string[] */
-    private $factories;
+    /** @var array<string, string> */
+    private array $invokables;
 
     /** @var FactoryInterface[]|callable[] */
-    private $resolvedFactories = [];
-
-    /** @var mixed[] */
-    private $resolvedServices = [];
-
-    /** @var mixed[] */
-    private $services;
+    private array $resolvedFactories = [];
+    private array $resolvedServices = [];
 
     /**
-     * @param mixed[]                                $services
      * @param FactoryInterface[]|callable[]|string[] $factories
      * @param AbstractFactoryInterface[]             $abstractFactories
+     * @param string[]                               $invokables
      */
-    public function __construct(array $services = [], array $factories = [], array $abstractFactories = [])
-    {
-        $this->abstractFactories = $abstractFactories;
-        $this->factories         = $factories;
-        $this->services          = $services;
+    public function __construct(
+        private array $services = [],
+        private array $factories = [],
+        private array $abstractFactories = [],
+        array $invokables = [],
+    ) {
+        $this->invokables = array_combine($invokables, $invokables);
     }
 
     public function addAbstractFactory(AbstractFactoryInterface $abstractFactory): void
@@ -54,44 +48,103 @@ class ServiceManager implements ArrayAccess, ContainerInterface
         $this->abstractFactories[] = $abstractFactory;
     }
 
-    public function addFactory(string $name, $factory): void
+    public function addFactory(string $id, string|callable $factory): void
     {
-        if (is_string($factory) || is_callable($factory)) {
-            $this->factories[$name] = $factory;
-        } else {
-            throw new TypeError(sprintf('Invalid factory for service %s given!', $name));
-        }
+        $this->factories[$id] = $factory;
     }
 
-    public function addService(string $name, $service): void
+    public function addInvokable(string $id): void
     {
-        $this->offsetSet($name, $service);
+        $this->invokables[$id] ??= $id;
+    }
+
+    public function addService(string $id, $service): void
+    {
+        $this->services[$id] = $service;
     }
 
     /**
      * @throws ContainerException
      */
-    public function createService(string $name, array $options = [])
+    public function createService(string $id, ?array $options = null)
     {
         try {
-            return $this->getFactory($name)($this, $name, $options);
+            return $this->getFactory($id)($this, $id, $options);
         } catch (Throwable $t) {
-            throw new ContainerException(sprintf('Service %s could not be created!', $name), 0, $t);
+            throw new ContainerException(sprintf('Service "%s" could not be created!', $id), 0, $t);
         }
     }
 
-    /**
-     * @inheritDoc
-     */
-    public function get($id)
+    public function get(string $id)
     {
-        return $this->offsetGet($id);
+        if (array_key_exists($id, $this->services)) {
+            return $this->services[$id];
+        }
+
+        if (array_key_exists($id, $this->resolvedServices)) {
+            return $this->resolvedServices[$id];
+        }
+
+        $service                     = $this->createService($id);
+        $this->resolvedServices[$id] = $service;
+
+        return $service;
     }
 
-    private function getAbstractFactory(string $name): ?callable
+    public function has(string $id): bool
+    {
+        if (array_key_exists($id, $this->services) || isset($this->factories[$id])) {
+            return true;
+        }
+
+        foreach ($this->abstractFactories as $abstractFactory) {
+            if ($abstractFactory->canCreate($this, $id)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    public function offsetExists(mixed $offset): bool
+    {
+        return $this->has($offset);
+    }
+
+    /**
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function offsetGet(mixed $offset): mixed
+    {
+        return $this->get($offset);
+    }
+
+    public function offsetSet(mixed $offset, mixed $value): void
+    {
+        $this->addService($offset, $value);
+    }
+
+    public function offsetUnset(mixed $offset): void
+    {
+        $this->removeService($offset);
+    }
+
+    public function removeService(string $id): void
+    {
+        unset(
+            $this->factories[$id],
+            $this->invokables[$id],
+            $this->resolvedFactories[$id],
+            $this->resolvedServices[$id],
+            $this->services[$id]
+        );
+    }
+
+    private function getAbstractFactory(string $id): ?AbstractFactoryInterface
     {
         foreach ($this->abstractFactories as $abstractFactory) {
-            if ($abstractFactory->canCreate($this, $name)) {
+            if ($abstractFactory->canCreate($this, $id)) {
                 return $abstractFactory;
             }
         }
@@ -100,23 +153,24 @@ class ServiceManager implements ArrayAccess, ContainerInterface
     }
 
     /**
-     * @return FactoryInterface|callable
      * @throws ContainerException
      */
-    private function getFactory(string $name): callable
+    private function getFactory(string $id): FactoryInterface|callable
     {
-        if (isset($this->resolvedFactories[$name])) {
-            return $this->resolvedFactories[$name];
+        if (isset($this->resolvedFactories[$id])) {
+            return $this->resolvedFactories[$id];
         }
 
-        $resolvableFactory = $this->factories[$name] ?? $this->getAbstractFactory($name);
+        $resolvableFactory = $this->factories[$id]
+            ?? $this->getInvokableFactory($id)
+            ?? $this->getAbstractFactory($id);
 
         if (empty($resolvableFactory)) {
-            throw new ContainerException(sprintf('Factory for service %s not found!', $name));
+            throw new ContainerException(sprintf('Factory for service "%s" not found!', $id));
         }
 
         if (is_callable($resolvableFactory)) {
-            $this->resolvedFactories[$name] = $resolvableFactory;
+            $this->resolvedFactories[$id] = $resolvableFactory;
 
             return $resolvableFactory;
         }
@@ -125,72 +179,19 @@ class ServiceManager implements ArrayAccess, ContainerInterface
             $factory = new $resolvableFactory();
 
             if ($factory instanceof FactoryInterface) {
-                $this->resolvedFactories[$name] = $factory;
+                $this->resolvedFactories[$id] = $factory;
 
                 return $factory;
             }
         }
 
-        throw new ContainerException(sprintf('Factory for service %s is invalid!', $name));
+        throw new ContainerException(sprintf('Factory for service "%s" is invalid!', $id));
     }
 
-    public function has($id): bool
+    private function getInvokableFactory(string $id): ?InvokableFactory
     {
-        return $this->offsetExists($id);
-    }
-
-    public function offsetExists($offset): bool
-    {
-        if (array_key_exists($offset, $this->services) || isset($this->factories[$offset])) {
-            return true;
-        }
-
-        foreach ($this->abstractFactories as $abstractFactory) {
-            if ($abstractFactory->canCreate($this, $offset)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @inheritDoc
-     * @throws ContainerException
-     */
-    public function offsetGet($offset)
-    {
-        if (array_key_exists($offset, $this->services)) {
-            return $this->services[$offset];
-        }
-
-        if (array_key_exists($offset, $this->resolvedServices)) {
-            return $this->resolvedServices[$offset];
-        }
-
-        $service                         = $this->createService($offset);
-        $this->resolvedServices[$offset] = $service;
-
-        return $service;
-    }
-
-    public function offsetSet($offset, $value): void
-    {
-        $this->services[$offset] = $value;
-    }
-
-    public function offsetUnset($offset): void
-    {
-        unset(
-            $this->factories[$offset],
-            $this->resolvedFactories[$offset],
-            $this->resolvedServices[$offset],
-            $this->services[$offset]
-        );
-    }
-
-    public function removeService(string $name): void
-    {
-        $this->offsetUnset($name);
+        return isset($this->invokables[$id])
+            ? new InvokableFactory()
+            : null;
     }
 }
