@@ -8,9 +8,8 @@ use BlackBonjour\ServiceManager\FactoryInterface;
 use Psr\Container\ContainerInterface;
 
 /**
- * @author    Erick Dyck <info@erickdyck.de>
- * @since     18.09.2019
- * @copyright Copyright (c) 2019 Erick Dyck
+ * @author Erick Dyck <info@erickdyck.de>
+ * @since  18.09.2019
  */
 interface AbstractFactoryInterface extends FactoryInterface
 {
