@@ -8,9 +8,8 @@ use Error;
 use Psr\Container\ContainerExceptionInterface;
 
 /**
- * @author    Erick Dyck <info@erickdyck.de>
- * @since     13.05.2019
- * @copyright Copyright (c) 2019 Erick Dyck
+ * @author Erick Dyck <info@erickdyck.de>
+ * @since  13.05.2019
  */
 class ContainerException extends Error implements ContainerExceptionInterface
 {
