@@ -7,18 +7,17 @@ namespace BlackBonjour\ServiceManager;
 use Psr\Container\ContainerInterface;
 
 /**
- * @author    Erick Dyck <info@erickdyck.de>
- * @since     13.05.2019
- * @copyright Copyright (c) 2019 Erick Dyck
+ * @author Erick Dyck <info@erickdyck.de>
+ * @since  13.05.2019
  */
 interface FactoryInterface
 {
     /**
      * Creates a new service.
      *
-     * @param ContainerInterface $container A container implementing PSR-11
-     * @param string             $service   Name of the service to create a new instance of
-     * @param array              $options   Some options that can be passed into the creation process
+     * @param ContainerInterface $container A container implementing PSR-11.
+     * @param string             $service   Name of the service to create a new instance of.
+     * @param array|null         $options   Some options that can be passed to build the service.
      */
-    public function __invoke(ContainerInterface $container, string $service, array $options = []);
+    public function __invoke(ContainerInterface $container, string $service, ?array $options = null);
 }
