@@ -5,25 +5,17 @@ declare(strict_types=1);
 namespace BlackBonjourTest\ServiceManager\AbstractFactory;
 
 use BlackBonjour\ServiceManager\AbstractFactory\DynamicFactory;
+use BlackBonjour\ServiceManager\Exception\ContainerException;
 use BlackBonjourTest\ServiceManager\Asset\FooBar;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 
 /**
- * @author    Erick Dyck <info@erickdyck.de>
- * @since     18.09.2019
- * @copyright Copyright (c) 2019 Erick Dyck
+ * @author Erick Dyck <info@erickdyck.de>
+ * @since  18.09.2019
  */
 class DynamicFactoryTest extends TestCase
 {
-    public function testInvoke(): void
-    {
-        self::assertInstanceOf(
-            FooBar::class,
-            (new DynamicFactory())($this->createMock(ContainerInterface::class), FooBar::class)
-        );
-    }
-
     public function testCanCreate(): void
     {
         $container = $this->createMock(ContainerInterface::class);
@@ -31,5 +23,24 @@ class DynamicFactoryTest extends TestCase
 
         self::assertTrue($factory->canCreate($container, FooBar::class));
         self::assertFalse($factory->canCreate($container, self::class));
+    }
+
+    public function testInvoke(): void
+    {
+        $container = $this->createMock(ContainerInterface::class);
+        $factory   = new DynamicFactory();
+
+        self::assertInstanceOf(FooBar::class, ($factory)($container, FooBar::class));
+    }
+
+    public function testInvokeFails(): void
+    {
+        $this->expectException(ContainerException::class);
+        $this->expectExceptionMessage(sprintf('Cannot create service "%s"!', self::class));
+
+        $container = $this->createMock(ContainerInterface::class);
+        $factory   = new DynamicFactory();
+
+        ($factory)($container, self::class);
     }
 }
