@@ -4,15 +4,11 @@ declare(strict_types=1);
 
 namespace BlackBonjourTest\ServiceManager\Asset;
 
-/**
- * @author Erick Dyck <info@erickdyck.de>
- * @since  18.09.2019
- */
-class FooBar
+class ClassWithoutFactoryAndScalarTypeHint
 {
     public function __construct(
+        public readonly int $id,
         public readonly string $foo,
-        public readonly string $bar,
     ) {
     }
 }
